+++ reddit_info_bot/__init__.py
@@ -29,82 +29,12 @@ from .search import (
     #get_karmadecay_results,
     get_tineye_results,
 )
-from .reddit import reddit_login
-from .antispam import spamfilter_lists
-from .util import domain_suffix, tld_from_suffix
-from .util import remove_control_characters
+from .reddit import reddit_login, reddit_msg_linkfilter
+from .antispam import spamfilter_lists, isspam
+from .util import domain_suffix, remove_control_characters
 
 logger = logging.getLogger(__name__)
 
-
-def isspam(result):
-    """check search result for spammy content
-    """
-    url, text = result[0].lower(), result[1].lower()
-
-    if len(url) < 6: # shorter than '//a.bc' can't be a useable absolute HTTP URL
-        print('Skipping invalid URL: "{0}"'.format(url))
-        return True
-    # domain from URL using publicsuffix (not a validator)
-    domain = domain_suffix(url)
-    if not domain:
-        print('Failed to lookup PSL/Domain for: "{0}"'.format(url))
-        return True
-    tld = tld_from_suffix(domain)
-    if not tld or tld == '':
-        print('Failed to lookup TLD from publicsuffix for: "{0}"'.format(url))
-        return True
-    if domain in whitelist:
-        # higher prio than the blacklist
-        return False
-    if domain in hard_blacklist:
-        print('Skipping blacklisted Domain "{0}": {1}'.format(domain, url))
-        return True
-    if tld in tld_blacklist:
-        print('Skipping blacklisted TLD "{0}": {1}'.format(tld, url))
-        return True
-    if url in link_filter:
-        print('Skipping spammy link match "{0}": {1}'.format(link_filter[url], url))
-        return True
-    if text in text_filter:
-        print('Skipping spammy text match "{0}": "{1}"'.format(text_filter[text], text))
-        return True
-    # no spam, result is good
-    return False
-
-def reddit_msg_linkfilter(messages, sending_account, receiving_account, submission_id):
-    """ Post reddit comments with one account and check from second account
-    to see if they were filtered out.
-    """
-    queue = []
-    submission = sending_account.get_submission(submission_id)
-    # post with first account
-    for message in messages:
-        try:
-            submission.add_comment(message)
-        except Exception as e:
-            print('reddit_msg_linkfilter failed to post "%s"' % (message,))
-            print(e)
-            # FIXME: check exception for http errors (retry?) or other (spam?)
-            continue
-        queue += [message]
-    time.sleep(7) # wait a bit
-    # fetch posts from second account
-    verified_messages = []
-    fetched_messages = receiving_account.get_unread(limit=40)
-    for msg in fetched_messages:
-        if msg.body in queue:
-            message = queue.pop(queue.index(msg.body))
-            msg.mark_as_read()
-            verified_messages += [message]
-        else:
-            print('reddit_msg_linkfilter skipping unknown message "%s"' % msg.body)
-    if queue: # messages posted but not found
-        print('reddit_msg_linkfilter posted but did not find: %s' % str(queue))
-    failed_messages = [m for m in messages if (m not in verified_messages and m not in queue)]
-    if failed_messages:
-        print('reddit_msg_linkfilter failed on: %s' % str(failed_messages))
-    return verified_messages
 
 def _reddit_spamfilter(results, sending_account, receiving_account, submission_id):
     urls = set([url for url, text in results])
@@ -115,13 +45,17 @@ def _reddit_spamfilter(results, sending_account, receiving_account, submission_i
             verified_results += result
     return verified_results
 
-def _format_results(results, display_limit=5): #returns a formatted and spam filtered list of the results. Change 5 to adjust number of results to display per provider. Fi
+def _format_results(results, display_limit=5):
+    """Format search results
+    Returns a markdown-formatted and spam-filtered list of the results.
+
+    `display_limit`: pass `None` to disable
+    """
     def sanitize_string(string):
         # strip possible control characters
         string = remove_control_characters(string)
 
-        # also strip non-ascii characters (old code)
-        # (not necessary, general unicode should be allowed)
+        # also strip non-ascii characters
         #string = ''.join(c for c in string if ord(c) in range(32, 127))
 
         string = string.strip()
@@ -160,7 +94,8 @@ def _format_results(results, display_limit=5): #returns a formatted and spam fil
                for result in results]
 
     # limit output to `display_limit` results
-    results = results[:display_limit]
+    if display_limit:
+        results = results[:display_limit]
 
     # format output
     markdown_links = ['[%s](%s)' % (text, url) for url, text in results]
@@ -185,12 +120,12 @@ def give_more_info(submission_url):
 
     # substitute videos with gif versions where possible
     # (because search engines index those)
-    domain = domain_suffix(url)
+    domain = domain_suffix(submission_url)
     if domain in ('imgur.com', 'gfycat.com'):
-        if submission_url.endswith(('.gifv', '.mp4', '.webm')):
-            submission_url = submission_url.replace('.mp4', '.gif')
-            submission_url = submission_url.replace('.gifv', '.gif')
-            submission_url = submission_url.replace('.webm', '.gif')
+        fileformats = ('.gifv', '.mp4', '.webm', '.ogg')
+        if submission_url.endswith(fileformats):
+            for ff in fileformats:
+                submission_url = submission_url.replace(ff, '.gif')
             print('Found %s video - substituting with gif url: %s' % (domain, submission_url))
         elif urlsplit(submission_url).path.rstrip(string.ascii_lowercase+string.ascii_uppercase) == '/':
             submission_url += '.gif'
@@ -263,8 +198,8 @@ def give_more_info(submission_url):
 # Bot actions
 #
 
-def reply_to_potential_comment(comment, attempt): #uncomment 'return true' to disable this feature
-    if (not config['USE_KEYWORDS']):
+def reply_to_potential_comment(comment, account):
+    if not config['USE_KEYWORDS']:
         return True
     if not any(i in str(comment.submission.url) for i in config['IMAGE_FORMATS']):
         return True
@@ -277,17 +212,14 @@ def reply_to_potential_comment(comment, attempt): #uncomment 'return true' to di
         elif botmode == LOG:
             print(reply)
         elif botmode == PM:
-             print(account1.send_message(comment.author, 'Info Bot Information', reply))
+             print(account.send_message(comment.author, 'Info Bot Information', reply))
         print("replied to potential comment: {0}".format(comment.body))
         done = True
         already_done.append(comment.id)
-    except requests.HTTPError:
+    except requests.HTTPError as e:
         done = True
-        print('HTTP Error. Bot might be banned from this sub')
+        print('HTTP Error. Bot might be banned from this sub:', e)
         already_done.append(comment.id)
-    except RateLimitExceeded:
-        print('submission rate exceeded! attempt %i'%attempt)
-        time.sleep(30)
     return done
 
 def find_username_mentions():
@@ -385,8 +317,6 @@ def find_keywords(all_comments):
         if not any(i for i in submission_comments if i.body == config['INFORMATION_REPLY']): #If there are information replies
             print('R', end='')
             continue
-        # previously determined already
-        #if any(word.lower() in comment.body.lower() for word in keyword_list):
         try:
             print("\ndetected keyword: "+ comment.body.lower())
         except UnicodeEncodeError:
@@ -404,12 +334,19 @@ def find_keywords(all_comments):
         done = False
         attempt = 1
         while not done:
-            done = reply_to_potential_comment(comment, attempt)
+            try:
+                done = reply_to_potential_comment(comment, account1)
+            except RateLimitExceeded as e:
+                print('submission rate exceeded! attempt %i' % attempt)
+                print(e)
+                time.sleep(30)
+
     #se = '/'.join(['%d %s' % (v, k) for k, v in stats])
     #print(' (%d comments - %s)' % (count, se))
     print(' (%d comments)' % (count,))
 
 def check_downvotes(user, start_time):
+    # FIXME: should check for comment's creation time
     current_time = int(time.time()/60)
     if (current_time - start_time) >= comment_deleting_wait_time:
         my_comments = user.get_comments(limit=None)
@@ -467,12 +404,10 @@ def main():
                 find_keywords(all_comments)
                 print('finding username mentions: ', end='')
                 find_username_mentions()
-                start_time = check_downvotes(user,start_time)
+                start_time = check_downvotes(user, start_time)
 
                 with open("already_done.p", "wb") as df:
                     pickle.dump(already_done, df)
-                with open("blacklist.p", "wb") as bf:
-                    pickle.dump(blacklist, bf)
 
                 print('Finished a round of comments. Waiting two seconds.\n')
                 time.sleep(2)
@@ -506,17 +441,6 @@ if __name__ == "__main__" or True: # always do this, for now
 
     time_limit_minutes = config['TIME_LIMIT_MINUTES'] #how long before a comment will be ignored for being too old
     comment_deleting_wait_time = config["DELETE_WAIT_TIME"] #how many minutes to wait before deleting downvoted comments
-
-    # load spam lists
-    (
-        link_filter,
-        text_filter,
-        word_filter,
-        hard_blacklist,
-        whitelist,
-        tld_blacklist,
-        blacklist,
-    ) = spamfilter_lists()
 
     already_done = []
     if os.path.isfile("already_done.p"):
